LibWeb: Move style update logic out of Document.cpp
Move the document style update algorithms into CSS/UpdateStyle.cpp and
expose them as CSS free functions operating on Document. Keep the
existing Document entry points as thin wrappers.

Add narrow Document hooks for the state the updater needs instead of
making the CSS functions friends of Document.
