+++ Libraries/LibWeb/DOM/Document.cpp
@@ -55,10 +55,8 @@
 #include <LibWeb/CSS/ComputedProperties.h>
 #include <LibWeb/CSS/FontComputer.h>
 #include <LibWeb/CSS/FontFaceSet.h>
-#include <LibWeb/CSS/Invalidation/HasMutationInvalidator.h>
 #include <LibWeb/CSS/Invalidation/MediaQueryInvalidator.h>
 #include <LibWeb/CSS/Invalidation/PseudoClassInvalidator.h>
-#include <LibWeb/CSS/Invalidation/SlotInvalidator.h>
 #include <LibWeb/CSS/Invalidation/StyleInvalidator.h>
 #include <LibWeb/CSS/InvalidationSet.h>
 #include <LibWeb/CSS/MediaQueryList.h>
@@ -66,7 +64,6 @@
 #include <LibWeb/CSS/Parser/Parser.h>
 #include <LibWeb/CSS/SelectorEngine.h>
 #include <LibWeb/CSS/StyleComputer.h>
-#include <LibWeb/CSS/StyleInvalidation.h>
 #include <LibWeb/CSS/StyleSheetIdentifier.h>
 #include <LibWeb/CSS/StyleSheetList.h>
 #include <LibWeb/CSS/StyleValues/ColorSchemeStyleValue.h>
@@ -2023,227 +2020,9 @@ bool Document::layout_is_up_to_date() const
         && m_svg_roots_needing_relayout.is_empty();
 }
 
-static void apply_element_style_invalidation_after_style_change(Element& element, CSS::RequiredInvalidationAfterStyleChange const& invalidation, bool invalidate_assigned_slottables, bool invalidate_descendant_slots)
+void Document::update_style_computer_viewport_rect()
 {
-    if (invalidate_assigned_slottables)
-        CSS::Invalidation::invalidate_assigned_slottables_after_slot_style_change(element);
-    if (invalidate_descendant_slots && (invalidation.inherited_style_changed || invalidation.rebuild_layout_tree))
-        CSS::Invalidation::invalidate_assigned_slottables_for_descendant_slots_after_inherited_style_change(element);
-
-    if (invalidation.relayout)
-        element.set_needs_layout_update(SetNeedsLayoutReason::StyleChange);
-    if (invalidation.rebuild_layout_tree) {
-        // Mark the parent to handle display changes to/from contents correctly.
-        if (auto parent_element = element.parent_element())
-            parent_element->set_needs_layout_tree_update(true, SetNeedsLayoutTreeUpdateReason::StyleChange);
-        else
-            element.set_needs_layout_tree_update(true, SetNeedsLayoutTreeUpdateReason::StyleChange);
-    }
-
-    element.set_needs_style_update(false);
-}
-
-static void apply_document_style_invalidation_after_style_change(Document& document, CSS::RequiredInvalidationAfterStyleChange const& invalidation)
-{
-    if (invalidation.repaint
-        || invalidation.rebuild_stacking_context_tree
-        || invalidation.relayout
-        || invalidation.rebuild_layout_tree)
-        document.set_needs_to_record_display_list();
-    if (invalidation.rebuild_accumulated_visual_contexts)
-        document.set_needs_accumulated_visual_contexts_update(true);
-    if (invalidation.rebuild_stacking_context_tree)
-        document.invalidate_stacking_context_tree();
-}
-
-[[nodiscard]] static CSS::RequiredInvalidationAfterStyleChange update_style_recursively(
-    Node& node,
-    CSS::StyleComputer& style_computer,
-    bool needs_inherited_style_update,
-    bool recompute_elements_depending_on_custom_properties,
-    bool parent_display_changed,
-    bool ancestor_needs_descendant_style_recompute)
-{
-    bool const needs_full_style_update = node.document().needs_full_style_update();
-    CSS::RequiredInvalidationAfterStyleChange invalidation;
-
-    if (node.is_element())
-        style_computer.push_ancestor(static_cast<Element const&>(node));
-
-    // NOTE: If the current node has `display:none`, we can disregard all invalidation
-    //       caused by its children, as they will not be rendered anyway.
-    //       We will still recompute style for the children, though.
-    bool is_display_none = false;
-
-    bool did_change_custom_properties = false;
-    CSS::RequiredInvalidationAfterStyleChange node_invalidation;
-    if (is<Element>(node)) {
-        auto& element = static_cast<Element&>(node);
-
-        // FIXME: We can avoid some unnecessary re-computations by, skipping if a) the contained if() functions don't
-        //        include media conditions, or b) the data used to resolve media queries hasn't changed.
-        bool const needs_style_update_due_to_if_media = element.style_uses_if_css_function();
-
-        if (needs_full_style_update
-            || node.needs_style_update()
-            || parent_display_changed
-            || ancestor_needs_descendant_style_recompute
-            || (recompute_elements_depending_on_custom_properties && (element.style_uses_var_css_function() || element.style_uses_inherit_css_function()))
-            || needs_style_update_due_to_if_media) {
-            node_invalidation = element.recompute_style(did_change_custom_properties);
-        } else {
-            if (needs_inherited_style_update)
-                node_invalidation = element.recompute_inherited_style(ScheduleAnimationUpdate::Yes);
-            if (recompute_elements_depending_on_custom_properties && element.refresh_inherited_custom_property_data())
-                did_change_custom_properties = true;
-        }
-        is_display_none = static_cast<Element&>(node).computed_properties()->display().is_none();
-
-        apply_element_style_invalidation_after_style_change(element, node_invalidation, !node_invalidation.is_none(), !needs_inherited_style_update);
-    }
-    if (!node.is_element())
-        node.set_needs_style_update(false);
-    invalidation |= node_invalidation;
-
-    if (did_change_custom_properties) {
-        recompute_elements_depending_on_custom_properties = true;
-    }
-
-    bool children_need_inherited_style_update = !invalidation.is_none();
-    if (!node.is_element())
-        children_need_inherited_style_update |= needs_inherited_style_update;
-    // NB: When display changes to/from flex/grid/contents, children may need to be blockified or un-blockified.
-    //     This requires a full style recompute, not just inherited style update.
-    bool children_need_full_style_recompute = node_invalidation.rebuild_layout_tree;
-    bool descendant_style_recompute_needed = ancestor_needs_descendant_style_recompute || node_invalidation.recompute_descendant_styles;
-
-    // OPTIMIZATION: Descendants of a display:none element are not rendered and their computed style is not observable
-    //               except through on-demand reads, which call Document::update_style_for_element to refresh the path
-    //               lazily. We can therefore skip the descent entirely. The exception is when display itself just
-    //               changed or the document needs a full style update. In those cases descendants must be re-cascaded
-    //               eagerly.
-    bool const skip_display_none_descent = is_display_none && !needs_full_style_update && !children_need_full_style_recompute;
-
-    if (!skip_display_none_descent
-        && (needs_full_style_update
-            || node.child_needs_style_update()
-            || children_need_inherited_style_update
-            || recompute_elements_depending_on_custom_properties
-            || children_need_full_style_recompute
-            || descendant_style_recompute_needed)) {
-        if (node.is_element()) {
-            if (auto shadow_root = static_cast<DOM::Element&>(node).shadow_root()) {
-                bool shadow_tree_children_need_inherited_style_update = node_invalidation.inherited_style_changed
-                    || (!node_invalidation.is_none() && shadow_root->children_may_depend_on_non_inherited_property_inheritance());
-                if (needs_full_style_update
-                    || shadow_root->needs_style_update()
-                    || shadow_root->child_needs_style_update()
-                    || shadow_tree_children_need_inherited_style_update
-                    || recompute_elements_depending_on_custom_properties
-                    || children_need_full_style_recompute
-                    || descendant_style_recompute_needed) {
-                    auto subtree_invalidation = update_style_recursively(
-                        *shadow_root,
-                        style_computer,
-                        shadow_tree_children_need_inherited_style_update,
-                        recompute_elements_depending_on_custom_properties,
-                        children_need_full_style_recompute,
-                        descendant_style_recompute_needed);
-                    if (!is_display_none)
-                        invalidation |= subtree_invalidation;
-                }
-            }
-        }
-
-        node.for_each_child([&](auto& child) {
-            if (needs_full_style_update
-                || child.needs_style_update()
-                || children_need_inherited_style_update
-                || child.child_needs_style_update()
-                || recompute_elements_depending_on_custom_properties
-                || children_need_full_style_recompute
-                || descendant_style_recompute_needed) {
-                auto subtree_invalidation = update_style_recursively(
-                    child,
-                    style_computer,
-                    children_need_inherited_style_update,
-                    recompute_elements_depending_on_custom_properties,
-                    children_need_full_style_recompute,
-                    descendant_style_recompute_needed);
-                if (!is_display_none)
-                    invalidation |= subtree_invalidation;
-            }
-            return IterationDecision::Continue;
-        });
-    }
-
-    if (!skip_display_none_descent)
-        node.set_child_needs_style_update(false);
-
-    if (node.is_element())
-        style_computer.pop_ancestor(static_cast<Element const&>(node));
-
-    return invalidation;
-}
-
-void Document::update_style()
-{
-    // NOTE: If our parent document needs a relayout, we must do that *first*. This is required as it may cause the
-    // viewport to change which will can affect media query evaluation and the value of the `vw` unit.
-    if (auto navigable = this->navigable(); navigable && navigable->container() && &navigable->container()->document() != this)
-        navigable->container()->document().update_layout(UpdateLayoutReason::ChildDocumentStyleUpdate);
-
-    if (!browsing_context())
-        return;
-
-    // Fetch the viewport rect once, instead of repeatedly, during style computation.
     style_computer().set_viewport_rect({}, viewport_rect());
-
-    update_animated_style_if_needed();
-
-    // Associated with each top-level browsing context is a current transition generation that is incremented on each
-    // style change event. [CSS-Transitions-2]
-    m_transition_generation++;
-
-    if (m_needs_invalidation_of_elements_affected_by_has) {
-        m_needs_invalidation_of_elements_affected_by_has = false;
-        CSS::Invalidation::invalidate_style_for_pending_has_mutations(*this);
-    }
-
-    if (!m_style_invalidator->has_pending_invalidations() && !needs_full_style_update() && !needs_style_update() && !child_needs_style_update() && !m_needs_media_rule_evaluation)
-        return;
-
-    // NOTE: If this is a document hosting <template> contents, style update is unnecessary.
-    if (m_created_for_appropriate_template_contents)
-        return;
-
-    if (m_needs_media_rule_evaluation)
-        evaluate_media_rules();
-
-    if (!m_style_invalidator->has_pending_invalidations() && !needs_full_style_update() && !needs_style_update() && !child_needs_style_update())
-        return;
-
-    m_style_invalidator->invalidate(*this);
-
-    build_registered_properties_cache();
-
-    CSS::RequiredInvalidationAfterStyleChange invalidation;
-    constexpr size_t max_style_update_passes = 8;
-    for (size_t style_update_pass = 0; style_update_pass < max_style_update_passes; ++style_update_pass) {
-        style_computer().reset_has_result_cache();
-        style_computer().reset_ancestor_filter();
-
-        invalidation |= update_style_recursively(*this, style_computer(), false, false, false, false);
-        m_needs_full_style_update = false;
-
-        if (!m_style_invalidator->has_pending_invalidations() && !needs_style_update() && !child_needs_style_update())
-            break;
-
-        m_style_invalidator->invalidate(*this);
-    }
-
-    apply_document_style_invalidation_after_style_change(*this, invalidation);
-    update_animated_style_if_needed();
 }
 
 static bool element_or_pseudo_depends_on_viewport_metrics(Element const& element)
@@ -2275,224 +2054,6 @@ void Document::invalidate_style_for_viewport_change()
 
         return TraversalDecision::Continue;
     });
-}
-
-void Document::update_style_if_needed_for_element(AbstractElement const& abstract_element)
-{
-    if (element_needs_style_update(abstract_element))
-        update_style();
-}
-
-static void mark_direct_children_for_style_update(Element& element)
-{
-    if (auto shadow_root = element.shadow_root()) {
-        shadow_root->for_each_child([](auto& child) {
-            child.set_needs_style_update(true);
-            return IterationDecision::Continue;
-        });
-    }
-
-    element.for_each_child([](auto& child) {
-        child.set_needs_style_update(true);
-        return IterationDecision::Continue;
-    });
-}
-
-static void apply_targeted_style_invalidation(Element& element, CSS::RequiredInvalidationAfterStyleChange const& invalidation, bool did_change_custom_properties, bool descendant_style_recompute_needed)
-{
-    apply_element_style_invalidation_after_style_change(element, invalidation, !invalidation.is_none() || did_change_custom_properties, true);
-
-    if (!invalidation.is_none() || did_change_custom_properties || descendant_style_recompute_needed || invalidation.recompute_descendant_styles)
-        mark_direct_children_for_style_update(element);
-
-    apply_document_style_invalidation_after_style_change(element.document(), invalidation);
-}
-
-static CSS::RequiredInvalidationAfterStyleChange recompute_style_for_targeted_style_update(Element& element, bool& did_change_custom_properties)
-{
-    if (element.parent())
-        return element.recompute_style(did_change_custom_properties);
-
-    auto new_computed_properties = element.document().style_computer().compute_style({ element }, did_change_custom_properties);
-    element.set_computed_properties({}, move(new_computed_properties));
-    element.set_needs_style_update(false);
-    return {};
-}
-
-CSS::ComputedProperties const* Document::update_style_for_element(AbstractElement const& abstract_element, StyleUpdateMode mode)
-{
-    // Refresh computed properties for an abstract element without requiring every unrelated dirty element in the
-    // document to be resolved. This walks the flat-tree inheritance chain and re-cascades from the rootmost stale
-    // element on the path back down to the target. Normal mode also re-cascades the target path under display:none
-    // ancestors, because the regular document traversal may leave those descendants stale.
-
-    if (auto navigable = this->navigable(); navigable && navigable->container() && &navigable->container()->document() != this)
-        navigable->container()->document().update_layout(UpdateLayoutReason::ChildDocumentStyleUpdate);
-
-    if (browsing_context()) {
-        style_computer().set_viewport_rect({}, viewport_rect());
-
-        update_animated_style_if_needed();
-
-        // Media query evaluation can enqueue normal style invalidations, so do it before deciding whether the full
-        // style traversal needs to run.
-        if (m_needs_media_rule_evaluation)
-            evaluate_media_rules();
-
-        if (!m_is_running_update_layout
-            && (needs_full_style_update()
-                || m_needs_invalidation_of_elements_affected_by_has
-                || m_style_invalidator->has_pending_invalidations()
-                || needs_style_update())) {
-            update_style();
-        }
-    }
-
-    // Single walk up the inheritance chain: collect each ancestor and remember the index of the topmost display:none
-    // entry seen. Pseudo-element styles are refreshed when the originating element is recomputed, so don't put the pseudo
-    // on the path.
-    GC::RootVector<GC::Ref<Element>> inheritance_chain;
-    if (!abstract_element.pseudo_element().has_value())
-        inheritance_chain.append(const_cast<Element&>(abstract_element.element()));
-
-    Optional<size_t> topmost_display_none_index;
-    Optional<size_t> topmost_element_requiring_style;
-    bool ancestor_needs_descendant_style_recompute = false;
-    for (auto cursor = abstract_element.element_to_inherit_style_from(); cursor.has_value(); cursor = cursor->element_to_inherit_style_from()) {
-        auto& ancestor = const_cast<Element&>(cursor->element());
-        inheritance_chain.append(ancestor);
-    }
-
-    for (size_t i = inheritance_chain.size(); i > 0; --i) {
-        auto& ancestor = inheritance_chain[i - 1];
-        if (!topmost_element_requiring_style.has_value()
-            && (ancestor_needs_descendant_style_recompute
-                || ancestor->needs_style_update()
-                || !ancestor->computed_properties())) {
-            topmost_element_requiring_style = i - 1;
-        }
-
-        if (ancestor->entire_subtree_needs_style_update()) {
-            if (!topmost_element_requiring_style.has_value())
-                topmost_element_requiring_style = i - 1;
-            ancestor_needs_descendant_style_recompute = true;
-        }
-
-        if (auto const properties = ancestor->computed_properties(); properties && properties->display().is_none()) {
-            topmost_display_none_index = i - 1;
-            if (mode == StyleUpdateMode::StopAtDisplayNone && !topmost_element_requiring_style.has_value())
-                return nullptr;
-        }
-    }
-
-    Optional<size_t> topmost_element_to_recompute = topmost_element_requiring_style;
-    if (mode == StyleUpdateMode::Normal && topmost_display_none_index.has_value()) {
-        if (!topmost_element_to_recompute.has_value() && *topmost_display_none_index > 0)
-            topmost_element_to_recompute = *topmost_display_none_index - 1;
-    }
-
-    if (!topmost_element_to_recompute.has_value()) {
-        if (mode == StyleUpdateMode::Normal && !inheritance_chain.is_empty())
-            topmost_element_to_recompute = 0;
-        else
-            return abstract_element.computed_properties();
-    }
-
-    style_computer().reset_has_result_cache();
-
-    // Re-cascading the inheritance chain requires the style computer's ancestor filter to reflect each recomputed
-    // element's DOM ancestors, so that descendant-combinator selectors match correctly. The filter is empty at this
-    // point because the normal top-down `update_style` traversal skipped the display:none subtree, so we have to seed
-    // it ourselves.
-    GC::RootVector<GC::Ref<Element>> ancestor_chain;
-    for (auto* cursor = inheritance_chain[*topmost_element_to_recompute]->parent_or_shadow_host_element(); cursor; cursor = cursor->parent_or_shadow_host_element())
-        ancestor_chain.append(*cursor);
-
-    auto& style_computer = this->style_computer();
-    for (size_t i = ancestor_chain.size(); i > 0; --i)
-        style_computer.push_ancestor(ancestor_chain[i - 1]);
-
-    GC::RootVector<GC::Ref<Element>> pushed_path_ancestors;
-
-    ScopeGuard pop_ancestor_chain = [&] {
-        for (auto& ancestor : ancestor_chain)
-            style_computer.pop_ancestor(ancestor);
-    };
-
-    ScopeGuard pop_path_ancestors = [&] {
-        for (auto& ancestor : pushed_path_ancestors)
-            style_computer.pop_ancestor(ancestor);
-    };
-
-    bool descendant_style_recompute_needed = false;
-    for (size_t i = *topmost_element_to_recompute + 1; i > 0; --i) {
-        auto& element = inheritance_chain[i - 1];
-        bool did_change_custom_properties = false;
-        auto invalidation = recompute_style_for_targeted_style_update(element, did_change_custom_properties);
-        apply_targeted_style_invalidation(element, invalidation, did_change_custom_properties, descendant_style_recompute_needed);
-
-        descendant_style_recompute_needed |= invalidation.recompute_descendant_styles;
-
-        if (element->computed_properties()->display().is_none()) {
-            if (mode == StyleUpdateMode::StopAtDisplayNone)
-                return nullptr;
-            descendant_style_recompute_needed = false;
-        }
-
-        if (did_change_custom_properties || invalidation.rebuild_layout_tree)
-            descendant_style_recompute_needed = true;
-
-        if (i > 1) {
-            style_computer.push_ancestor(element);
-            pushed_path_ancestors.append(element);
-        }
-    }
-
-    return abstract_element.computed_properties();
-}
-
-bool Document::element_needs_style_update(AbstractElement const& abstract_element) const
-{
-    // If there are document-level reasons to update style, we can't skip.
-    if (m_needs_full_style_update)
-        return true;
-    if (m_needs_animated_style_update)
-        return true;
-    if (m_needs_invalidation_of_elements_affected_by_has)
-        return true;
-    if (m_needs_media_rule_evaluation)
-        return true;
-    if (m_style_invalidator->has_pending_invalidations())
-        return true;
-
-    // Check the element itself.
-    if (abstract_element.element().needs_style_update())
-        return true;
-    if (abstract_element.element().entire_subtree_needs_style_update())
-        return true;
-
-    // Walk the inheritance ancestor chain. We use element_to_inherit_style_from()
-    // because style inheritance follows the flat tree (slotted elements inherit
-    // from their assigned slot, not their DOM parent). If any ancestor on the
-    // path has its style marked dirty, the target element's computed style could
-    // change via inherited properties, so we must update.
-    for (auto ancestor = abstract_element.element_to_inherit_style_from(); ancestor.has_value(); ancestor = ancestor->element_to_inherit_style_from()) {
-        if (ancestor->element().needs_style_update())
-            return true;
-        if (ancestor->element().entire_subtree_needs_style_update())
-            return true;
-    }
-
-    // If the navigable has a container and that container needs a style update, then we need one as well, since the
-    // container's style can affect this element (e.g. via media queries or viewport units).
-    if (auto navigable = this->navigable()) {
-        if (auto container = navigable->container()) {
-            if (container->document().element_needs_style_update(AbstractElement { *container }))
-                return true;
-        }
-    }
-
-    return false;
 }
 
 void Document::update_animated_style_if_needed()
